% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{automate_keywords}
\alias{automate_keywords}
\title{Automated super function}
\usage{
automate_keywords(
  seedWords,
  corpus,
  modelPath,
  minSentenceChar = 3,
  minTermFrequency = 2,
  nVectors = 5000,
  nCandidates = 200,
  weightSimilarity = 2,
  keepPOS = c("NOUN", "ADJ", "X", "PROPN", "ENTITY", "nounphrase", "PRON")
)
}
\arguments{
\item{seedWords}{starting seed words}

\item{corpus}{domain relevant training text corpus}

\item{modelPath}{path to pre-trained word2vec embedding text}
}
\value{
list of keywords
}
\description{
One function to clean, process, and generate new keywords
}
