% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{clean_corpus}
\alias{clean_corpus}
\title{Clean training text corpus}
\usage{
clean_corpus(seedWords, corpus, minSentenceChar = 3)
}
\arguments{
\item{seedWords}{starting search string keywords.}

\item{corpus}{quanteda formatted text corpus.}

\item{minSentenceChar}{minimum number of words contained in a sentence.}
}
\value{
sentence level quanteda corpus.
}
\description{
Clean training text corpus
}
