% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{clean_vocabulary}
\alias{clean_vocabulary}
\title{Further clean word list using rules.}
\usage{
clean_vocabulary(wordList)
}
\arguments{
\item{wordList}{domain relevant vacabulary.}
}
\value{
cleaned vocabulary.
}
\description{
Further clean word list using rules.
}
