% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{convert_bool}
\alias{convert_bool}
\title{Convert Regex or Boolean query to string}
\usage{
convert_bool(string, regexPattern = " or | and | doc |\\\\*")
}
\arguments{
\item{string}{regex or boolean style query}
}
\value{
list of keywords
}
\description{
Convert Regex or Boolean query to string
}
