% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{create_query}
\alias{create_query}
\title{Export keyword results to formatted query}
\usage{
create_query(results, n = 25, type = c("boolean", "regex"))
}
\arguments{
\item{results}{list of keywords}

\item{n}{number of keywords in new query}

\item{type}{boolean or regex style query}
}
\value{
paste friendly search strings
}
\description{
Export keyword results to formatted query
}
