% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{export_query}
\alias{export_query}
\title{Format search query}
\usage{
export_query(results, type = c("boolean", "regex"))
}
\arguments{
\item{results}{list of keywords}

\item{type}{boolean or regex style query}
}
\value{
cleaned search query
}
\description{
Format search query
}
