% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{get_centroid}
\alias{get_centroid}
\title{Get average embedding for a list of words based on centroid.}
\usage{
get_centroid(dictionary, vectors)
}
\arguments{
\item{dictionary}{list of domain relevant words}

\item{vectors}{word embeddings}
}
\value{
normalized centroids for word vectors
}
\description{
Get average embedding for a list of words based on centroid.
}
