% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{get_keywords}
\alias{get_keywords}
\title{Get word recommendations most similar to starting search string.}
\usage{
get_keywords(dictionary, vectors)
}
\arguments{
\item{dictionary}{list of domain relevant words}

\item{vectors}{word embeddings}
}
\value{
cosine similarity
}
\description{
Get word recommendations most similar to starting search string.
}
