% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Get filteres list of unique corpus domain relevant words.}
\usage{
get_vocabulary(parsedText, minTermFrequency = 2)
}
\arguments{
\item{parsedText}{fully cleaned corpus with tagging and filtering.}

\item{minTermFrequency}{minimum number of times a word appears in a text to be kept.}
}
\value{
unique list of words in corpus.
}
\description{
Get filteres list of unique corpus domain relevant words.
}
