% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{load_glove}
\alias{load_glove}
\title{Filter domain relevant vocabulary from pre-trained GloVe embeddings.}
\usage{
load_glove(filename, corpus, n = 5000, minTermFrequency = 2)
}
\arguments{
\item{filename}{path to pre-trained GloVe word embedding text file.}

\item{corpus}{cleaned text corpus}

\item{n}{total top words to keep from cleaned text corpus}

\item{minTermFrequency}{minimum occurances of word in training text}
}
\value{
list of embeddings relevant to domain vocabulary.
}
\description{
Filter domain relevant vocabulary from pre-trained GloVe embeddings.
}
