% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{make_bool}
\alias{make_bool}
\title{Convert Regex to Boolean}
\usage{
make_bool(regexQ)
}
\arguments{
\item{regexQ}{Regex style query}
}
\value{
Boolean style query
}
\description{
Convert Regex to Boolean
}
