% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{multi_word_avg}
\alias{multi_word_avg}
\title{Average embeddings for multi-part words}
\usage{
multi_word_avg(vocabulary, vectors)
}
\arguments{
\item{vocabulary}{words}

\item{vectors}{embeddings}
}
\value{
average embedding for multi-part words.
}
\description{
Average embeddings for multi-part words
}
