% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_funct.R
\name{tag_corpus}
\alias{tag_corpus}
\title{Parse and tag corpus with entities and parts of speech}
\usage{
tag_corpus(tidyCorp)
}
\arguments{
\item{tidyCorp}{corpus returned by clean_corpus}
}
\value{
parsed and lemmatised text with entity and pos tags
}
\description{
Parse and tag corpus with entities and parts of speech
}
